+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Changelog
 ### Version 1.0.8
-* 
+* Fixed crash when changing the preview background color
 ### Version 1.0.7
 * Adding images now inserts the new images after the currently selected image
 * Added user configurable FPS setting